requirements.txt
langchain_core>=0.1.4
langchain_openai>=0.0.2
langchain_chroma>=0.0.4
langgraph>=0.0.15
chromadb>=0.4.18
openai>=1.6.0
python-dotenv>=1.0.0
requests>=2.31.0
typing_extensions>=4.8.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
tldextract>=3.4.0
validators>=0.22.0
rich>=13.5.0
environs>=9.5.0